+++ groups/feeds.py
@@ -8,7 +8,7 @@ from reader.models import Chapter
 
 from .models import Group
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from datetime import datetime
     from django.http import HttpRequest